refactored robot state publishing and logging
