+++ src/main/java/org/team5924/frc2025/RobotState.java
@@ -20,8 +20,8 @@ import edu.wpi.first.math.geometry.Pose2d;
 import lombok.Getter;
 import lombok.Setter;
 import org.littletonrobotics.junction.AutoLogOutput;
-import org.team5924.frc2025.subsystems.algae.AlgaePivot.AlgaePivotState;
 import org.team5924.frc2025.subsystems.elevator.Elevator.ElevatorState;
+import org.team5924.frc2025.subsystems.pivot.AlgaePivot.AlgaePivotState;
 import org.team5924.frc2025.subsystems.rollers.CoralInAndOut.CoralInAndOut.CoralState;
 
 @Getter
@@ -40,19 +40,11 @@ public class RobotState {
   @AutoLogOutput(key = "RobotState/EstimatedPose")
   private Pose2d estimatedPose = new Pose2d();
 
-  @Getter
-  @Setter
-  private ElevatorState elevatorState = ElevatorState.MANUAL;
+  @Getter @Setter private ElevatorState elevatorState = ElevatorState.MANUAL;
 
   /* ### Coral In and Out ### */
-  @Getter
-  @Setter
-  @AutoLogOutput(key = "RobotState/CoralState")
-  private CoralState coralInAndOutState = CoralState.NO_CORAL;
+  @Getter @Setter private CoralState coralInAndOutState = CoralState.NO_CORAL;
 
   /* ### Algae Pivot ### */
-  @Getter
-  @Setter
-  @AutoLogOutput(key = "RobotState/AlgaePivotState")
-  private AlgaePivotState algaePivotState = AlgaePivotState.INTAKE_FLOOR;
+  @Getter @Setter private AlgaePivotState algaePivotState = AlgaePivotState.INTAKE_FLOOR;
 }
