+++ src/main/java/org/team5924/frc2025/subsystems/pivot/AlgaePivot.java
@@ -14,7 +14,7 @@
  * If you did not, see <https://www.gnu.org/licenses>.
  */
 
-package org.team5924.frc2025.subsystems.algae;
+package org.team5924.frc2025.subsystems.pivot;
 
 import edu.wpi.first.wpilibj.Alert;
 import edu.wpi.first.wpilibj.DriverStation;
