+++ src/main/java/org/team5924/frc2025/subsystems/rollers/CoralInAndOut/CoralInAndOut.java
@@ -20,6 +20,7 @@ import java.util.function.DoubleSupplier;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import org.littletonrobotics.junction.Logger;
 import org.team5924.frc2025.RobotState;
 import org.team5924.frc2025.subsystems.rollers.GenericRollerSystem;
 import org.team5924.frc2025.subsystems.rollers.GenericRollerSystem.VoltageState;
@@ -73,6 +74,11 @@ public class CoralInAndOut extends GenericRollerSystem<CoralInAndOut.CoralState>
             goalState.getVoltageSupplier().getAsDouble(),
             goalState.getHandoffVoltage().getAsDouble());
     super.periodic();
+
+    Logger.recordOutput(
+        "RobotState/Algae/PivotState", RobotState.getInstance().getAlgaePivotState());
+    Logger.recordOutput(
+        "RobotState/Coral/InAndOutState", RobotState.getInstance().getCoralInAndOutState());
   }
 
   public void setGoalState(CoralState goalState) {
